Change all id's of domains from long to class Long to repair serializing. ServiceDAO now take data from spring application.

+++ src/main/java/ru/kafpin/lr6_bzz/dao/AutomobileDao.java
@@ -84,7 +84,7 @@ public class AutomobileDao implements Dao<Automobile, Long> {
         try {
             while (rs.next()){
                 list.add(new Automobile(
-                        rs.getInt("id"),
+                        rs.getLong("id"),
                         rs.getString("mark"),
                         rs.getString("model"),
                         rs.getString("gosnumber"),
@@ -179,7 +179,7 @@ public class AutomobileDao implements Dao<Automobile, Long> {
         try {
             while (rs.next()){
                 automobile = new Automobile(
-                        rs.getInt("id"),
+                        rs.getLong("id"),
                         rs.getString("mark"),
                         rs.getString("model"),
                         rs.getString("gosnumber"),
